fix - export TRPCError
