+++ packages/server/src/internals/errors.ts
@@ -1,29 +1,4 @@
-import { TRPC_ERROR_CODE_KEY } from './rpc/codes';
-
-export class TRPCError extends Error {
-  public readonly originalError?: unknown;
-  public readonly code;
-
-  constructor({
-    message,
-    code,
-    originalError,
-  }: {
-    message?: string;
-    code: TRPC_ERROR_CODE_KEY;
-    originalError?: unknown;
-  }) {
-    super(message ?? getMessageFromUnkownError(originalError, code));
-    this.code = code;
-    this.originalError = originalError;
-    this.name = 'TRPCError';
-
-    Object.setPrototypeOf(this, new.target.prototype);
-  }
-}
-export interface TRPCErrorOptions {
-  originalError?: unknown;
-}
+import { TRPCError } from '../TRPCError';
 
 export function getMessageFromUnkownError(
   err: unknown,
