+++ packages/server/src/router.ts
@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import { assertNotBrowser } from './assertNotBrowser';
-import { TRPCError } from './errors';
+import { TRPCError } from './TRPCError';
 import {
   createProcedure,
   CreateProcedureOptions,
