+++ packages/server/test/errors.test.ts
@@ -3,7 +3,8 @@ import { z, ZodError } from 'zod';
 import { TRPCClientError } from '../../client/src';
 import * as trpc from '../src';
 import { AnyRouter } from '../src';
-import { getMessageFromUnkownError, TRPCError } from '../src/errors';
+import { getMessageFromUnkownError } from '../src/internals/errors';
+import { TRPCError } from '../src/TRPCError';
 import { routerToServerAndClient } from './_testHelpers';
 
 function assertClientError(
